refactor: clean up imports and remove unused code in various test files
